Menu
I added some labels to Menu and I added some code to track when the view controller comes up (when you switch tabs). I also made an Evaluation class, which I think, when completed, will be able to do the work of both Poll and Image. I don't really like the name of Image because it's too close to the names of UIImage, UIImageView, imageview.image, etc. which gets confusing.